refactor: clear all lint warnings (type-safety + module splits)
Resolves the 16 standing ESLint warnings repo-wide; lint now reports
0 warnings / 0 errors. No behavior change.

Type-safety / dead-code fixes:
- Centralize the bridge's PendingExport<any> map into a single
  PendingExportMap alias (one documented suppression, was three)
- Validate untrusted localStorage in customBinRegistry instead of a
  blind double-cast; drop malformed entries individually
- Remove 4 dead type assertions (lid builders + a test)
- Drop a redundant `?? false` in FloatingInspector
- Re-type thumbnail.ts vertex-count guards (three.js types getAttribute
  as non-null but it returns undefined at runtime — the guard is real)
- Annotate the compartmentTexts index read in designerValidation as
  unknown, matching the file's existing cells convention

Module splits (under the 500-line max-lines limit, all acyclic):
- bin-designer types/index.ts -> cutout.ts
- constants/defaults.ts -> paramMigration.ts (deps injected, no cycle)
- BaseplatePanel.tsx -> CornerRadiusControl + GridDimensionStepper (+tests)
- api designerValidation.ts -> designerValidationConstants + compartment validators

The two print-critical geometry modules (compartments.ts, cutoutBuilder.ts)
get a justified file-level max-lines override rather than a risky
restructure of working boolean/grid code.
